fix: linter errors
